Wrap if statement around split() method
If the 'properties' for a given row is empty, the variable will be nil and the properties.split(":") method should not be executed - doing so causes an error when using the script with a CSV with empty 'properties'. There should be a check to ensure the 'properties' variable is populated before executing, otherwise the properties_to_hash() method should just return empty property_hash.